change database update strategy for account reset cancellation
