+++ spec/services/account_reset/pending_request_for_user_spec.rb
@@ -1,7 +1,7 @@
 require 'rails_helper'
 
-RSpec.describe AccountReset::FindPendingRequestForUser do
-  describe '#call' do
+RSpec.describe AccountReset::PendingRequestForUser do
+  describe '#get_account_reset_request' do
     let(:user) { create(:user) }
     let(:granted_at) { nil }
     let(:cancelled_at) { nil }
@@ -19,31 +19,31 @@ RSpec.describe AccountReset::FindPendingRequestForUser do
     subject { described_class.new(user) }
 
     context 'when a request exists' do
-      it { expect(subject.call).to eq(account_reset_request) }
+      it { expect(subject.get_account_reset_request).to eq(account_reset_request) }
     end
 
     context 'when a request does not exist' do
       let!(:account_reset_request) { nil }
 
-      it { expect(subject.call).to be_nil }
+      it { expect(subject.get_account_reset_request).to be_nil }
     end
 
     context 'when a request exists, but it has been granted' do
       let(:granted_at) { 1.hour.ago }
 
-      it { expect(subject.call).to be_nil }
+      it { expect(subject.get_account_reset_request).to be_nil }
     end
 
     context 'when a request exists, but it is expired' do
       let(:requested_at) { 1.year.ago }
 
-      it { expect(subject.call).to be_nil }
+      it { expect(subject.get_account_reset_request).to be_nil }
     end
 
     context 'when a request exists, but it has been cancelled' do
       let(:cancelled_at) { 1.hour.ago }
 
-      it { expect(subject.call).to be_nil }
+      it { expect(subject.get_account_reset_request).to be_nil }
     end
 
     context 'fraud user' do
@@ -52,12 +52,12 @@ RSpec.describe AccountReset::FindPendingRequestForUser do
       context 'when a request exists, and it hasnt been granted yet but over a day' do
         let(:requested_at) { 30.hours.ago }
 
-        it { expect(subject.call).to eq(account_reset_request) }
+        it { expect(subject.get_account_reset_request).to eq(account_reset_request) }
       end
       context 'when a request has expired' do
         let(:requested_at) { 1.year.ago }
 
-        it { expect(subject.call).to be_nil }
+        it { expect(subject.get_account_reset_request).to be_nil }
       end
     end
   end
